requirements.txt
streamlit
requests
beautifulsoup4
textblob
python-telegram-bot==13.15
pandas
